requirements.txt
streamlit
pyrebase4
python-dotenv
streamlit-option-menu
plotly